Fix line endings for the rest of the repository
I am aware that line endings don't matter for extlibs, but now that they
should be normalized automatically (see 96a8e724), git will pester
everybody to death if we leave them as is. Now this is settled once and
for all, and no line endings should annoy us ever again.


Former-commit-id: 665d46f770edec98562d799e6768c6738884c597